CMD_WRITE handled only in user context
